Allow nested loops
Also fix CT_LOOP_BREAK for good.

When a loop iterates for the first time, Globals.loopArgs is initialized with
the loop initial arguments. The loop commands are then put on the stack,
followed by a new CT_REGALIA_LOOPARGS command and the loop condition.
The CT_REGALIA_LOOPARGS command restores the properties of Globals.loopArgs
so that nested loops can change and use them too. This commands also set the
Globals.loopArgsValid flag so that the loop condition "knows" this is not the
first loop iteration.

The value of Globals.loopArgs.object is saved and restored when entering and
leaving a loop so that this property is always valid.